requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
sentence-transformers>=3.2.0,<4.0.0
faiss-cpu>=1.9.0,<2.0.0
numpy>=1.26.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
spacy>=3.8.0,<4.0.0
rank-bm25>=0.2.2,<1.0.0
groq>=0.11.0,<1.0.0
httpx>=0.27.0,<1.0.0
pytest>=8.3.0,<9.0.0
